+++ domain/src/main/java/com/amsterdam/domain/useCase/mood/GetMoviesByMoodUseCase.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.domain.useCase.home
+package com.amsterdam.domain.useCase.mood
 
 import com.amsterdam.domain.utils.Mood
 import com.amsterdam.domain.repository.MovieRepository
