+++ domain/src/main/java/com/amsterdam/domain/useCase/popular/GetPopularMoviesUseCase.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.domain.useCase.home
+package com.amsterdam.domain.useCase.popular
 
 import com.amsterdam.domain.repository.MovieRepository
 import com.amsterdam.entity.Movie
