+++ domain/src/main/java/com/amsterdam/domain/useCase/popular/GetPopularTvShowsUseCase.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.domain.useCase.home
+package com.amsterdam.domain.useCase.popular
 
 import com.amsterdam.domain.repository.TvShowRepository
 import com.amsterdam.entity.TvShow
