+++ domain/src/main/java/com/amsterdam/domain/useCase/topRated/GetTopRatedDataUseCase.kt
@@ -1,17 +1,17 @@
-package com.amsterdam.domain.useCase.home
+package com.amsterdam.domain.useCase.topRated
 
 
 import com.amsterdam.entity.Movie
 import com.amsterdam.entity.TvShow
 
-class GetTopRatedScreenDataUseCase(
+class GetTopRatedDataUseCase(
     private val getTopRatedMoviesUseCase: GetTopRatedMoviesUseCase,
     private val getTopRatedTvShowsUseCase: GetTopRatedTvShowsUseCase,
 ) {
 
     suspend operator fun invoke(
         page: Int = 1,
-    ) : TopRatedScreenData{
+    ) : TopRatedScreenData {
         return TopRatedScreenData(
             topRatedMovies = getTopRatedMoviesUseCase(page),
             topRatedTvShows = getTopRatedTvShowsUseCase(page),
