+++ domain/src/main/java/com/amsterdam/domain/useCase/topRated/GetTopRatedMoviesUseCase.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.domain.useCase.home
+package com.amsterdam.domain.useCase.topRated
 
 import com.amsterdam.domain.repository.MovieRepository
 import com.amsterdam.entity.Movie
