+++ domain/src/main/java/com/amsterdam/domain/useCase/topRated/GetTopRatedTvShowsUseCase.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.domain.useCase.home
+package com.amsterdam.domain.useCase.topRated
 
 import com.amsterdam.domain.repository.TvShowRepository
 import com.amsterdam.entity.TvShow
