+++ domain/src/test/java/com/amsterdam/domain/useCase/continueWatching/GetContinueWatchingDataUseCaseTest.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.domain.useCase.home
+package com.amsterdam.domain.useCase.continueWatching
 
 import com.amsterdam.domain.exceptions.AflamiException
 import com.amsterdam.domain.useCase.utils.fakeMovieList
