+++ viewModel/src/main/java/com/amsterdam/viewmodel/watchHistory/WatchHistoryViewModel.kt
@@ -8,8 +8,8 @@ import androidx.paging.cachedIn
 import androidx.paging.map
 import com.amsterdam.domain.exceptions.AflamiException
 import com.amsterdam.domain.exceptions.NoInternetException
-import com.amsterdam.domain.useCase.home.GetContinueWatchingMoviesUseCase
-import com.amsterdam.domain.useCase.home.GetContinueWatchingTvShowsUseCase
+import com.amsterdam.domain.useCase.continueWatching.GetContinueWatchingMoviesUseCase
+import com.amsterdam.domain.useCase.continueWatching.GetContinueWatchingTvShowsUseCase
 import com.amsterdam.domain.useCase.preferences.ManageLocaleLanguageUseCase
 import com.amsterdam.paging.PagingSource
 import com.amsterdam.viewmodel.shared.BaseViewModel
